+++ Sources/StreamVideo/WebRTC/v2/WebRTCStateAdapter.swift
@@ -138,7 +138,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate, W
         callCid: String,
         videoConfig: VideoConfig,
         callSettings: CallSettings,
-        clientEventReporter: ClientEventReporting = NoOpClientEventReporter(),
+        clientEventReporter: ClientEventReporting,
         rtcPeerConnectionCoordinatorFactory: RTCPeerConnectionCoordinatorProviding,
         stagePublisher: AnyPublisher<WebRTCCoordinator.StateMachine.Stage.ID, Never>,
         videoCaptureSessionProvider: VideoCaptureSessionProvider = .init(),
@@ -186,7 +186,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate, W
         videoConfig: VideoConfig,
         callSettings: CallSettings,
         peerConnectionFactory: PeerConnectionFactory,
-        clientEventReporter: ClientEventReporting = NoOpClientEventReporter(),
+        clientEventReporter: ClientEventReporting,
         rtcPeerConnectionCoordinatorFactory: RTCPeerConnectionCoordinatorProviding,
         stagePublisher: AnyPublisher<WebRTCCoordinator.StateMachine.Stage.ID, Never>,
         videoCaptureSessionProvider: VideoCaptureSessionProvider = .init(),
