+++ StreamVideoTests/WebRTC/SFU/SFUEventAdapter_Tests.swift
@@ -28,6 +28,7 @@ final class SFUEventAdapter_Tests: XCTestCase, @unchecked Sendable {
         callCid: .unique,
         videoConfig: Self.videoConfig,
         callSettings: .default,
+        clientEventReporter: NoOpClientEventReporter(),
         rtcPeerConnectionCoordinatorFactory: MockRTCPeerConnectionCoordinatorFactory(),
         stagePublisher: stageSubject.eraseToAnyPublisher()
     )
